refactor(common): add downloadAsFile helper and streamline download logic
- Introduced `downloadAsFile` in `components/common.tsx` to encapsulate blob creation and download.
- Updated `Sidebar.tsx` and `Settings.tsx` to use this helper instead of inline download code.
- Removes duplicated download logic and simplifies maintenance.
