+++ src/components/common.tsx
@@ -60,6 +60,20 @@ export const OpenInNewTab = ({
   </a>
 );
 
+export const downloadAsFile = (blobParts: BlobPart[], fileName: string) => {
+  const blob = new Blob(blobParts, {
+    type: 'application/json',
+  });
+  const url = URL.createObjectURL(blob);
+  const a = document.createElement('a');
+  a.href = url;
+  a.download = fileName;
+  document.body.appendChild(a);
+  a.click();
+  document.body.removeChild(a);
+  URL.revokeObjectURL(url);
+};
+
 /**
  * @deprecated Use `title` and `aria-label` props in button directly as utiliy classes.
  *
